--- src/targets/gpu/include/migraphx/gpu/device/rnn_last_output.hpp
@@ -1,24 +1,0 @@
-#ifndef MIGRAPHX_GUARD_RTGLIB_DEVICE_RNN_LAST_OUTPUT_HPP
-#define MIGRAPHX_GUARD_RTGLIB_DEVICE_RNN_LAST_OUTPUT_HPP
-
-#include <migraphx/argument.hpp>
-#include <migraphx/config.hpp>
-#include <hip/hip_runtime_api.h>
-
-namespace migraphx {
-inline namespace MIGRAPHX_INLINE_NS {
-namespace gpu {
-namespace device {
-
-void rnn_last_output(hipStream_t stream,
-                     const argument& result,
-                     const argument& arg_hs,
-                     const argument& arg_sl,
-                     bool is_reverse);
-
-} // namespace device
-} // namespace gpu
-} // namespace MIGRAPHX_INLINE_NS
-} // namespace migraphx
-
-#endif
