+++ src/targets/gpu/include/migraphx/gpu/device/rnn_variable_seq_lens.hpp
@@ -10,16 +10,22 @@ inline namespace MIGRAPHX_INLINE_NS {
 namespace gpu {
 namespace device {
 
+void rnn_var_sl_shift_sequence(hipStream_t stream,
+                               const argument& result,
+                               const argument& arg_hs,
+                               const argument& arg_sl);
+
 void rnn_var_sl_shift_output(hipStream_t stream,
                              const argument& result,
                              const argument& arg_hs,
                              const argument& arg_sl,
                              bool is_reverse);
 
-void rnn_var_sl_shift_sequence(hipStream_t stream,
-                               const argument& result,
-                               const argument& arg_hs,
-                               const argument& arg_sl);
+void rnn_var_sl_last_output(hipStream_t stream,
+                     const argument& result,
+                     const argument& arg_hs,
+                     const argument& arg_sl,
+                     bool is_reverse);
 
 } // namespace device
 } // namespace gpu
